Improve layout on small screens even more
Add a dropdown menu and move the hint link to it. When the screen is too
small to fit all the region buttons those now move to the menu as well.

Change the hover style of the region buttons to use a border instead of
a text drop shadow because that looked ugly.

Switch to a consistent font for everything and use system fonts.